DocSummarizer/my_agent/agent_server/requirements.txt
mlflow[databricks]>=3.1.3
langgraph>=0.3.0
langchain>=1.0.0
langchain-core>=0.3.0
databricks-langchain>=0.5.0
databricks-sdk>=0.50.0
pydantic>=2.0.0
flask>=3.0.0
asgiref>=3.7.0
pymupdf4llm>=0.0.17
python-dotenv